contracts-bedrock: fix flakes attempt
The problem is the address of the contract that does the deployments
of the proxies. When running as `forge script`, a different address is
the deployer based on when running as a test. The deployer contract is
etched into a particular address when running in the tests but this is
not the case when doing doing `forge script`.

This is an attempt to fix the flakes at the expense of devex. The
genesis generation script is now required to also build the L1 genesis
state. The L1 genesis state is exported using `vm.dumpState` and then
imported by the tests.

Note that this creates pain when developing the smart contracts
that should be alleviated when we move to solidity L2 genesis generation.